% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{add_last_finite_rotations}
\alias{add_last_finite_rotations}
\title{Last finite rotation in sequence}
\usage{
add_last_finite_rotations(x)
}
\arguments{
\item{x}{data.frame. Sequence of total reconstruction rotations}
}
\value{
data.frame
}
\description{
Adds last finite rotation in a sequence assuming that the plate motion does not change
}
\examples{
data(pangea)
add_last_finite_rotations(pangea)
}
