% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{antipodal_euler_pole}
\alias{antipodal_euler_pole}
\title{Antipodal Euler pole}
\usage{
antipodal_euler_pole(ep)
}
\arguments{
\item{ep}{Euler pole object}
}
\value{
Object of class \code{"euler.pole"}
}
\description{
Creates a Euler pole antidpodal location
}
\examples{
euler.pole <- euler_pole(90, 0)
antipodal_euler_pole(euler.pole)
}
