% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{antipodal_rotation}
\alias{antipodal_rotation}
\title{Antipodal rotation}
\usage{
antipodal_rotation(x)
}
\arguments{
\item{x}{data.frame containing the sequence of rotations or the rotation}
}
\value{
object of with same class like x
}
\description{
Euler pole on the other side of the hemisphere
}
\examples{
euler.pole <- data.frame(lat=c(27, -21), lon=c(17, -151), angle=c(0.4, 0.42))
antipodal_rotation(euler.pole)
}
