% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{as.finite}
\alias{as.finite}
\title{Finite rotation class}
\usage{
as.finite(x)
}
\arguments{
\item{x}{data.frame containing columns lat, lon, angle, plate.rot, and plate.fix
\describe{
  \item{plate.rot}{ID of moving plate}
  \item{age}{Age for finite rotation}
  \item{lat}{Latitude of Euler pole of total reconstruction rotation}
  \item{lon}{Longitude}
  \item{angle}{Rotation angle in degree}
  \item{plate.fix}{ID of fixed/anchored plate}
  }}
}
\value{
object of class \code{"finite"}
}
\description{
Assign a data.frame to class \code{'finite'}
}
