% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{best_cone}
\alias{best_cone}
\title{The cone of best fit to conically disposed s-plane poles}
\usage{
best_cone(x)
}
\arguments{
\item{x}{data.frame containing the three dimensions of coordinates:
\describe{
\item{\code{a}}{angular deviation from E-W (x)}
\item{\code{b}}{angular deviation from N-S (y)}
\item{\code{c}}{angular deviation from vertical (z)}
}}
}
\description{
Finding the best fit pole of rotation for a given set of points that are
assumed to lie on a mutual small circle
}
\examples{
x <- rbind(
  c(-67, -31, -71),
  c(-62, -53, -50),
  c(-62, -75, -34),
  c(-58, 85, -34),
  c(-79, 40, -52),
  c(90, 14, -75),
  c(80, 10, 90)
  ) \%>\% data.frame()
  colnames(x) <- c('a', 'b', 'c')
  best_cone(x)
  best_plane(x)
}
