% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{cartesian_to_spherical}
\alias{cartesian_to_spherical}
\title{Spherical coordinates}
\usage{
cartesian_to_spherical(x)
}
\arguments{
\item{x}{matrix}
}
\description{
Spherical coordinates
}
\examples{
x <- rbind(
  c(58, 78, 35),
  c(47, 72, 47),
  c(57, 68, 41),
  c(56, 64, 45),
  c(52, 65, 47),
  c(61, 62, 51),
  c(55, 60, 49),
  c(61, 62, 42),
  c(68, 52, 46),
  c(42, 58, 65)
)
spherical_to_cartesian(x)
cartesian_to_spherical(spherical_to_cartesian(x))
}
