% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{check.stage}
\alias{check.stage}
\title{Stage rotation object}
\usage{
check.stage(x)
}
\arguments{
\item{x}{data.frame containing columns lat, lon, angle, plate.rot, and plate.fix
\describe{
  \item{plate.rot}{ID of moving plate}
  \item{lat}{Latitude of Euler pole of total reconstruction rotation}
  \item{lon}{Longitude}
  \item{min.age}{End of stage rotation}
  \item{max.age}{Start of stage rotation}
  \item{angle}{Rotation angle in degree}
  \item{plate.fix}{ID of fixed/anchored plate}
  }}
}
\value{
data.frame
}
\description{
Check if object has columns of a 'stage' rotation sequence
}
