% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{clip_eulercircles}
\alias{clip_eulercircles}
\title{Clip Euler circles}
\usage{
clip_eulercircles(x, crop_by)
}
\arguments{
\item{x}{Euler circle object. Either an object of \code{"sf"} or
\code{"SpatialLinesDataFrame"}}

\item{crop_by}{an object of \code{"SpatialLinesDataFrame"}, \code{"sf"} or a
\code{"data.frame"} with the range of longitudes and latitudes (lon, lat).}
}
\value{
An object of \code{class(x)}
}
\description{
Clip small circles, great circles, or loxodromes at plate boundaries
}
