% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{clip_grid_by_plate}
\alias{clip_grid_by_plate}
\title{Clip grid along plate boundaries}
\usage{
clip_grid_by_plate(grid, plate)
}
\arguments{
\item{grid}{\code{"data.frame"}}

\item{plate}{an object of \code{"SpatialLinesDataFrame"} or \code{"sf"}}
}
\value{
\code{"data.frame"}
}
\description{
Crops a grid along plate boundaries
}
