% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{equivalent_rotation}
\alias{equivalent_rotation}
\title{Equivalent rotation}
\usage{
equivalent_rotation(x, fixed)
}
\arguments{
\item{x}{sequence of plate rotations. Object of class \code{"finite"} of \code{"stage"}}

\item{fixed}{ID of new fixed plate. Has to be one out of \code{x$plate.fix}}
}
\value{
sequence of plate rotations in new reference system. Same object class as x
}
\description{
Transform a sequence of rotations into a new reference system
}
\seealso{
\code{\link{'as.finite'}}, \code{\link{'as.stage'}}
}
