% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{equivalent_rotations}
\alias{equivalent_rotations}
\title{Equivalent rotations of different reference system}
\usage{
equivalent_rotations(x, fixed)
}
\arguments{
\item{x}{sequence of plate rotations. An object of \code{"finite"} of \code{"stage"}}

\item{fixed}{ID of fixed plate. Must be same object class as x}
}
\value{
sequence of plate rotations. Same object class as x
}
\description{
Transform a sequence of rotations with different reference
systems into one with a common reference system
}
