% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{euler_from_rot}
\alias{euler_from_rot}
\title{Euler axis and angle from Euler matrix}
\usage{
euler_from_rot(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
list
\describe{
\item{pole}{Euler pole object}
\item{psi}{Euler angle in degree}
}
}
\description{
Extracts the Euler pole and angle from a Euler matrix
}
