% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{euler_rot}
\alias{euler_rot}
\title{Euler rotation matrix}
\usage{
euler_rot(ep, psi)
}
\arguments{
\item{ep}{Euler pole object}

\item{psi}{Angle of rotation in degree}
}
\value{
\code{matrix}
}
\description{
Creates a matrix from the given set of values.
}
\examples{
euler.pole <- euler_pole(90, 0)
euler_rot(euler.pole, psi = 45)
}
\references{
<div class="csl-entry">Greiner, B. (1999). Euler rotations in plate-tectonic reconstructions. <i>Computers and Geosciences</i>, <i>25</i>(3), 209–216. https://doi.org/10.1016/S0098-3004(98)00160-5</div>
}
