% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{euler_rotation}
\alias{euler_rotation}
\title{Euler rotation}
\usage{
euler_rotation(ep, psi, x)
}
\arguments{
\item{ep}{Euler pole object}

\item{psi}{Angle of rotation in degree}

\item{x}{Point to be rotated. Two-column Vector with  latitude and longitude}
}
\value{
latitude and longitude of rotated vector
}
\description{
Performs a Euler rotation of a vector
}
\examples{
euler.pole <- euler_pole(90, 0)
euler_rotation(euler.pole, psi = 45, x = c(45, 45))
}
