% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{extract_stage_rotation_matrix}
\alias{extract_stage_rotation_matrix}
\title{Stage rotation extraction of rotation matrices}
\usage{
extract_stage_rotation_matrix(a1, a2)
}
\arguments{
\item{a1, a2}{rotation matrix of finite rotations (t(a2)<t(a1))}
}
\value{
list
}
\description{
extract stage rotation of two finite rotation matrices
}
\details{
x must be all equivalent total rotations.
}
\references{
Greiner, B. (1999). Euler rotations in plate-tectonic reconstructions. Computers and Geosciences, 25(3), 209–216. https://doi.org/10.1016/S0098-3004(98)00160-5
}
\seealso{
\code{\link{extract_stage_rotations}}, \code{\link{extract_stage_rotation}}
}
