% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{extract_stage_rotations}
\alias{extract_stage_rotations}
\title{Stage rotations extraction}
\usage{
extract_stage_rotations(x, plate)
}
\arguments{
\item{x}{object of class \code{"finite"}. Sequence of total
reconstruction rotations}

\item{plate}{???}
}
\value{
object of class \code{"stage"}. Sequence of stage rotations
}
\description{
extract all stage rotations from a sequence of  total
reconstruction rotations
}
\details{
x must ba all equivalent total rotations.
}
\references{
<div class="csl-entry">Greiner, B. (1999). Euler rotations in plate-tectonic reconstructions. <i>Computers and Geosciences</i>, <i>25</i>(3), 209–216. https://doi.org/10.1016/S0098-3004(98)00160-5</div>
}
