% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{find_missing_rotations}
\alias{find_missing_rotations}
\title{Find missing rotations in a rotatin sequence}
\usage{
find_missing_rotations(x)
}
\arguments{
\item{x}{data.frame. Sequence of total
reconstruction rotations}
}
\value{
data.frame. Sequence of total reconstruction rotations
}
\description{
Identifies gaps in a sequence of rotations with
different reference systems
}
\details{
x must ba all equivalent total rotations.
}
\seealso{
\code{\link{check.finite}}, \code{\link{check.stage}}
}
