% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gplates_ids}
\alias{gplates_ids}
\title{ID list of plates used in GPLATES}
\format{
An object of class \code{"tibble"}
}
\usage{
data('gplates_ids')
}
\description{
ID list of plates used in GPLATES
}
\examples{
data("gplates_ids")
}
\keyword{datasets}
