% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gplates_rot}
\alias{gplates_rot}
\title{Rotation parameters of GPlates}
\format{
An object of class \code{"tibble"}
}
\usage{
data('gplates_rot')
}
\description{
Rotation parameters of GPlates
}
\examples{
data("gplates_rot")
}
\keyword{datasets}
