% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{greatcircle_distance}
\alias{greatcircle_distance}
\title{Great-circle distance distance}
\usage{
greatcircle_distance(a, b)
}
\arguments{
\item{a}{lat, lon coordinates of point 1}

\item{b}{lat, lon coordinates of point 2}
}
\value{
distance in degree
}
\description{
Distance along a great circle.
}
\examples{
berlin <- c(52.517, 13.4)
tokyo <- c(35.7, 139.767)
greatcircle_distance(berlin, tokyo)
}
\seealso{
\code{\link{orthodrome}}
}
