% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greiner}
\alias{greiner}
\title{The rotations opening the Central and Northern Atlantic Ocean}
\format{
An object of class \code{"tibble"}
}
\usage{
data('greiner')
}
\description{
Total reconstruction rotations of Greiner and Neugebauer (2013)
}
\examples{
data("greiner")
}
\keyword{datasets}
