% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{interpolate_azimuth}
\alias{interpolate_azimuth}
\title{Inverse Distance Weighting with Directional Data}
\usage{
interpolate_azimuth(values, locations, x.range, y.range, spacing = 1, idp = 2)
}
\arguments{
\item{values}{A table of points to be interpolated. Table must contain x and y locations, and a column of values to be interpolated.}

\item{locations}{A table with lat and lon coordinates of the samples.}

\item{x.range}{Latitude range of the grid}

\item{y.range}{Longitude range of the grid}

\item{spacing}{Spacing of the grid. default is 1}

\item{idp}{The power to use in weight calculation. default is 2}
}
\description{
Function for inverse distance weighted interpolation with directional data. Useful for when you are working with data whose unit of measurement is degrees (i.e. the average of 35 degrees and 355 degrees should be 15 degrees). It works by finding the shortest distance between two degree marks on a circle.
}
