% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{interpolate_idw}
\alias{interpolate_idw}
\title{Inverse Distance Weighting interpolation}
\usage{
interpolate_idw(values, locations, x.range, y.range, spacing, idp, ...)
}
\arguments{
\item{values}{A table of points to be interpolated. Table must contain x and y locations, and a column of values to be interpolated.}

\item{locations}{A table with lat and lon coordinates of the samples.}

\item{x.range}{Latitude range of the grid}

\item{y.range}{Longitude range of the grid}

\item{spacing}{Spacing of the grid}

\item{idp}{The power to use in weight calculation.}

\item{...}{Other arguments to be passed to \code{\link[gstat]{idw}}}
}
\description{
This function interpolates a list of samples with location and a value to a table of coordinates, that generally represent a spatial grid. The interpolation is based on inverse distance weighting algoritm with three different methods available for weight calculation.
}
