% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{interpolate_missing_finite_poles}
\alias{interpolate_missing_finite_poles}
\title{Interpolate gaps in the sequence of total reconstruction rotations}
\usage{
interpolate_missing_finite_poles(df)
}
\arguments{
\item{df}{data.frame. Sequence of total
reconstruction rotations}
}
\value{
data.frame. Sequence of total
reconstruction rotations with filled gaps
}
\description{
Interpolate missing rotations in a sequence of rotations with a
different reference systems
}
\examples{
data(pangea)
interpolate_missing_finite_poles(pangea)
}
\seealso{
\code{\link{"check.finite"}}, \code{\link{check.stage}}
}
