% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{is_spatial}
\alias{is_spatial}
\alias{is.sp}
\alias{is.sf}
\title{Function to test if an R object is a Spatial* or a sf data-type.}
\usage{
is.sp(x)

is.sf(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
returns \code{TRUE} if its argument is a Spatial* or a sf
 (that is, has \code{"Spatial*"} or \code{"sf"} amongst its classes) and
 FALSE otherwise.
}
\description{
Helper functions to
}
