% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kroner_plates}
\alias{kroner_plates}
\title{Plates of the Earth for paleogeographical reconstructions for the
Pannotia-Pangea cycle}
\format{
An object of class \code{"sf"}. MULTIPOLYGON
}
\usage{
data('kroner_plates')
}
\description{
Global set of plates for Pangea-Pannotia supercontinent cycle
}
\examples{
data("kroner_plates")
}
\references{
<div class="csl-entry">Kroner, U., Stephan, T., Romer, R. L., &#38; Roscher, M. (2020). Paleozoic plate kinematics during the Pannotia–Pangaea supercontinent cycle. <i>Geological Society, London, Special Publications</i>, <i>503</i>, SP503-2020–15. https://doi.org/10.1144/SP503-2020-15</div>
}
\keyword{datasets}
