% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{mean_pole}
\alias{mean_pole}
\title{Unimodal pole distribution}
\usage{
mean_pole(x)
}
\arguments{
\item{x}{data.frame containing the three dimensions of coordinates:
\describe{
\item{\code{a}}{angular deviation from E-W (x)}
\item{\code{b}}{angular deviation from N-S (y)}
\item{\code{c}}{angular deviation from vertical (z)}
}}
}
\description{
mean vector, variance and standard deviation
}
\examples{
x <- rbind(
c(58, 78, 35),
c(47, 72, 47),
c(57, 68, 41),
c(56, 64, 45),
c(52, 65, 47),
c(61, 62, 51),
c(55, 60, 49),
c(61, 62, 42),
c(68, 52, 46),
c(42, 58, 65)
) \%>\% data.frame()
colnames(x) <- c("a", "b", "c")
mean_pole(x)
var_pole(x)
sd_pole(x)
}
