% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{misfit_euler_from_azimuths}
\alias{misfit_euler_from_azimuths}
\title{Misfit of Euler pole from azimuths of small-circle structures}
\usage{
misfit_euler_from_azimuths(lat, lon, azi, ep_lat, ep_lon)
}
\arguments{
\item{lat, lon, azi}{Latitude, longitude and azimuth of the small-circle structures (e.g. fracture zones along the crest of the sea-floor ridges), numeric in degrees}

\item{ep_lat, ep_lon}{estimated position of the Euler pole for relative plate motion, numeric in degrees}
}
\description{
Misfit of Euler pole from azimuths of small-circle structures
}
\details{
The actual search is made through a  grid centered on the original ep_lat and ep_lon
The process is repeated until the point is located to better than 1 degree.
}
\references{
Le Pichon, X. (1968), Sea-floor spreading and continental drift, J. Geophys. Res., 73( 12), 3661-3697, doi:10.1029/JB073i012p03661.
}
