% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{misfit_euler_from_rates}
\alias{misfit_euler_from_rates}
\title{Misfit of Euler pole from spreading rates}
\usage{
misfit_euler_from_rates(lat, lon, azi, v, ep_lat, ep_lon)
}
\arguments{
\item{lat, lon, azi}{Latitude, longitude and azimuth of the small-circle structures (e.g. fracture zones along the crest of the sea-floor ridges), numeric in degrees}

\item{v}{spreading rate measured along a perpendicular to the crest of the ridge}

\item{ep_lat, ep_lon}{estimated position of the Euler pole for relative plate motion, numeric in degrees}
}
\description{
Misfit of Euler pole from spreading rates
}
\references{
Le Pichon, X. (1968), Sea-floor spreading and continental drift, J. Geophys. Res., 73( 12), 3661-3697, doi:10.1029/JB073i012p03661.
}
