% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalization of a vector#'}
\usage{
normalize_vector(v)
}
\arguments{
\item{v}{vector}
}
\value{
normalized vector
}
\description{
normalizes a vector to unit length#'
}
\examples{
normalize_vector(1:5)
}
