% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pangea}
\alias{pangea}
\title{Example rotation file}
\format{
An object of class \code{"finite"}
}
\usage{
data(pangea)
}
\description{
Sequence of total reconstruction rotations in GPLATES' .rot format
}
\examples{
data(pangea)
head(pangea)
}
\references{
<div class="csl-entry">Kroner, U., Roscher, M., &#38; Romer, R. L. (2016). Ancient plate kinematics derived from the deformation pattern of continental crust: Paleo- and Neo-Tethys opening coeval with prolonged Gondwana–Laurussia convergence. <i>Tectonophysics</i>, <i>681</i>, 220–233. https://doi.org/10.1016/j.tecto.2016.03.034</div>
}
\keyword{datasets}
