% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pannotia}
\alias{pannotia}
\title{Example rotation file}
\format{
An object of class \code{"finite"}
}
\usage{
data(pannotia)
}
\description{
Sequence of total reconstruction reotations in GPLATE's .rot format
}
\examples{
data(pannotia)
head(pannotia)
}
\references{
<div class="csl-entry">Kroner, U., Stephan, T., Romer, R. L., &#38; Roscher, M. (2020). Paleozoic plate kinematics during the Pannotia–Pangaea supercontinent cycle. <i>Geological Society, London, Special Publications</i>, <i>503</i>, SP503-2020–15. https://doi.org/10.1144/SP503-2020-15</div>
}
\keyword{datasets}
