% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pb2002_boundaries}
\alias{pb2002_boundaries}
\title{Plate Boundaries of the Earth}
\format{
An object of class \code{"sf"}. LINESTRING
}
\usage{
data('pb2002_boundaries')
}
\description{
Global set of present plate boundaries on the Earth (PB2002) by Bird 2003
}
\examples{
data("pb2002_boundaries")
head("pb2002_boundaries")
}
\references{
Bird, P. (2003), An updated digital model of plate boundaries,
*Geochem. Geophys. Geosyst.*, 4, 1027, doi: 10.1029/2001GC000252, 3.
}
\keyword{datasets}
