% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{plate_motion_grid}
\alias{plate_motion_grid}
\title{Plate motion grid}
\usage{
plate_motion_grid(
  euler,
  gridsize = 5,
  lat.lim = c(-90, 90),
  lon.lim = c(-180, 180)
)
}
\arguments{
\item{euler}{data.frame. containing lat, lon, and angle (optional) of Euler rotation}

\item{gridsize}{grid size in degree}

\item{lat.lim}{vector, range of latitudes}

\item{lon.lim}{vector, range of longitudes}
}
\value{
data.frame with plate motion direction (degree) and velocity (km/Myr)
at grid point
}
\description{
Create a grid of the plate motion direction and velocity
}
\examples{
data(pangea)
euler <- subset(pangea, pangea$plate.rot == 103 & pangea$age == 250)
plate_motion_grid(euler)
}
