% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{pole_distribution}
\alias{pole_distribution}
\title{Position statistics}
\usage{
pole_distribution(x)
}
\arguments{
\item{x}{a `matrix` containing the coordinates of various positions}
}
\description{
Statistics on the distribution of geographic locations
}
