% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{ramsay_coords}
\alias{ramsay_coords}
\alias{cartesian_to_acoscartesian}
\alias{acoscartesian_to_cartesian}
\alias{geographical_to_cartesian2}
\alias{cartesian_to_geographical2}
\alias{geographical_to_acoscartesian}
\alias{acoscartesian_to_geographical}
\title{Spherical coordinate conversion}
\usage{
cartesian_to_acoscartesian(x)

acoscartesian_to_cartesian(x)

geographical_to_cartesian2(x)

cartesian_to_geographical2(x)

geographical_to_acoscartesian(x)

acoscartesian_to_geographical(x)
}
\arguments{
\item{x}{matrix of  coordinates}
}
\description{
Spherical coordinate conversion
}
\examples{
# Stereographic coordinates (angle notation):
x <- rbind(
  c(58, 78, 35),
  c(47, 72, 47),
  c(57, 68, 41),
  c(56, 64, 45),
  c(52, 65, 47),
  c(61, 62, 51),
  c(55, 60, 49),
  c(61, 62, 42),
  c(68, 52, 46),
  c(42, 58, 65)
)
acoscartesian_to_cartesian(x)
acoscartesian_to_geographical(x)
acoscartesian_to_cartesian(x) |> cartesian_to_acoscartesian()
}
