% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{rose2}
\alias{rose2}
\title{Rose Diagram}
\usage{
rose2(
  x,
  bins = NULL,
  axial = FALSE,
  clockwise = TRUE,
  start = "N",
  unit = c("degree", "radian"),
  sub,
  ...
)
}
\arguments{
\item{x}{Data to be plotted. A numeric vector containing angles, or a
histogram object containing a histogram of angular values, or a density
object containing a smooth density estimate for angular data, or an fv object
giving a function of an angular argument.}

\item{bins}{number of arcs to partition the circle with.}

\item{axial}{Logical value indicating whether data are uniaxial (axial=FALSE)
or biaxial (axial=TRUE, the default).}

\item{clockwise}{Logical value indicating whether angles increase in the
clockwise direction (clockwise=TRUE, the default) or anti-clockwise,
counter-clockwise direction (clockwise=FALSE).}

\item{unit}{The unit in which the angles are expressed. "degree" the default,
or "radian".}

\item{...}{Additional arguments passed to [spatstat.explore::rose()].}
}
\value{
A window (class "owin") containing the plotted region.
}
\description{
Plots a rose diagram (rose of directions), the analogue of a
histogram or density plot for angular data.
}
\examples{
x <- runif(100, 60, 210)
rose2(x, col = "grey", axial = TRUE)
}
\author{
Copyright (C) 2021 Tobias Stephan
}
