% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotation}
\alias{rotation}
\title{Rotation}
\usage{
rotation(x, n, alpha)
}
\arguments{
\item{x}{vector in cartesian coordinates}

\item{n}{rotation axis in cartesian coordinates}

\item{alpha}{angle in degree}
}
\value{
rotated vector in cartesian coordinates
}
\description{
Rotates a vector around an axis and an angle
}
\details{
Rotation of a vector is the dot product of the rotation matrix and
the vector
}
