% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotation_angle}
\alias{rotation_angle}
\title{Rotation angle from rotation matrix}
\usage{
rotation_angle(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
angle in degree
}
\description{
Extracts the rotation anglr from rotation matrix
}
