% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotation_axis}
\alias{rotation_axis}
\title{Rotation axis from rotation matrix}
\usage{
rotation_axis(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
vector
}
\description{
Extracts the rotation axis from rotation matrix
}
