% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation matrix}
\usage{
rotation_matrix(n, alpha)
}
\arguments{
\item{n}{vector in cartesian coordinates}

\item{alpha}{in degree}
}
\value{
3x3 matrix
}
\description{
Creates rotation matrix from a rotation axis and angle
}
