% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seton_rot}
\alias{seton_rot}
\title{Rotation parameters from Seton et al. 2012}
\format{
An object of class \code{"tibble"}
}
\usage{
data('seton_rot')
}
\description{
Rotation parameters from Seton et al. 2012
}
\examples{
data("seton_rot")
}
\keyword{datasets}
