% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{sf.as.data.frame}
\alias{sf.as.data.frame}
\title{Simple feature to data.frame}
\usage{
sf.as.data.frame(x)
}
\arguments{
\item{x}{an object of class \code{sf}}
}
\value{
an object of class \code{data.frame}
}
\description{
Converts an \code{sf} object into a \code{data.frame}
}
