% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{spherical_angle}
\alias{spherical_angle}
\alias{orthodrome_haversine2}
\alias{orthodrome_haversine2_2}
\alias{orthodrome_vincenty2}
\title{Angle along great circle on spherical surface}
\usage{
orthodrome_haversine2(p1, p2)

orthodrome_haversine2_2(p1, p2)

orthodrome_vincenty2(p1, p2)
}
\arguments{
\item{p1, p2}{numeric vector; lat, lon coordinates of point p1 and p2}
}
\value{
angle in radians
}
\description{
Smallest angle between two points on the surface of a sphere, measured along
the surface of the sphere
}
\details{
\describe{
\item{"orthodrome_haversine"}{uses Haversine formula (the default)}
\item{"orthodrome_haversine2"}{ uses Haversine formula optimized for 64-bit floating-point numbers}
\item{"orthodrome_vincenty"}{uses Vincenty formula for an ellipsoid with equal major and minor axes}
}
}
\examples{
berlin <- c(52.52, 13.41)
calgary <- c(51.04, -114.072)
orthodrome_haversine2(berlin, calgary)
orthodrome_haversine2_2(berlin, calgary)
orthodrome_vincenty2(berlin, calgary)
}
