% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{spherical_distance}
\alias{spherical_distance}
\alias{dist_greatcircle2}
\alias{dist_loxodrome}
\alias{dist_smallcircle}
\title{Distances on Earth' surface}
\usage{
dist_greatcircle2(
  p1,
  p2,
  r = earth_radius(),
  method = c("haversine", "haversine2", "vincenty")
)

dist_loxodrome(p1, p2, r = earth_radius())

dist_smallcircle(
  p1,
  p2,
  sm,
  r = earth_radius(),
  method = c("haversine", "haversine2", "vincenty")
)
}
\arguments{
\item{p1, p2}{numeric vector; lat, lon coordinates of point p1 and p2}

\item{r}{radius of the sphere (default = 6371.0087714 km, i.e. the radius of
the Earth)}

\item{method}{Formula for calculating great circle distance:
\describe{
\item{"haversine"}{Haversine formula (the default)}
\item{"haversine2"}{Haversine formula optimized for 64-bit floating-point numbers}
\item{"vincenty"}{Vincenty formula for an ellipsoid with equal major and minor axes}
}}

\item{sm}{Angle between pole and small circle.}
}
\value{
distance in units of r (default = kilometers)
}
\description{
Distances on a sphere about the size of the Earth.
}
\details{
The great-circle distance or orthodromic distance is the
shortest distance between two points on the surface of a sphere,
measured along the surface of the sphere.

The loxodromic distance between two points gives the distance measured along
a line of constant bearing.

The small circle distance between two points is measured along a small circle
given by the angle between the pole and the small circle.
}
\note{
The Earth is nearly spherical, so spherical formulas give the distance
between points on the surface of the Earth correct to within about 0.5%.
}
\examples{
berlin <- c(52.52, 13.41)
calgary <- c(51.04, -114.072)
dist_greatcircle2(berlin, calgary)
dist_loxodrome(berlin, calgary)
dist_smallcircle(berlin, calgary, sm=45)
}
\references{
Imboden, C., Imboden D. (1972) Orthodromic and loxodromic formula for the
calculation of distance and direction between ringing and finding place.
Vogelwarte 26: 336-346.
}
