% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_rotationR.R
\name{stress2grid}
\alias{stress2grid}
\title{Stress2Grid}
\usage{
stress2grid(
  azi,
  unc,
  type,
  lon,
  lat,
  lon.range,
  lat.range,
  gridsize = 2.5,
  min_data = 3,
  threshold = 25,
  arte_thres = 200,
  apply_mw = FALSE,
  apply_qw = TRUE,
  dist_weight = c("linear", "inverse", "none"),
  dist_threshold = 0.1,
  R_range = seq(50, 1000, 50)
)
}
\arguments{
\item{azi}{SHmax in degree}

\item{unc}{Uncertainties of SHmax in degree}

\item{type}{methods used for the determination of the orientation of SHmax}

\item{gridsize}{Specify the spacing of the regular grid in decimal degree.}

\item{min_data}{minimum number of data per bin. Default is 3}

\item{threshold}{threshold for deviation of orientation. Default is 25}

\item{arte_thres}{Maximum distance (in km) of the gridpoint to the next datapoint. Default is 200}

\item{apply_mw}{Logical. If a method weighting should be applied: Default is FALSE}

\item{apply_qw}{Logical. If a quality weighting should be applied: Default is TRUE}

\item{dist_weight}{Distance weighting method which should be used. Specify ‘none’, ‘linear’, or ‘inverse’}

\item{dist_threshold}{distance weight to prevent overweight of data nearby (0 to 1). Default is 0.1}
}
\description{
Stress pattern and wavelength analysis
}
