% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{to_geomat}
\alias{to_geomat}
\title{Convert object to geographic coordinate matrix}
\usage{
to_geomat(x)
}
\arguments{
\item{x}{a `matrix` or `data.frame` containing the geographical coordinates,
or a `sf` object}
}
\value{
matrix
}
\description{
Convert object to geographic coordinate matrix
}
