% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Torsvik_APWP}
\alias{Torsvik_APWP}
\title{Torsvik et al. 2012 APWP}
\format{
An object of class \code{"sf"}. MULTIPOLYGON
}
\usage{
data('Torsvik_APWP')
}
\description{
Torsvik et al. 2012 APWP in African coordinates
}
\examples{
data("Torsvik_APWP")
}
\keyword{datasets}
