% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{antipodal_euler_pole}
\alias{antipodal_euler_pole}
\title{Antipodal Euler pole}
\usage{
antipodal_euler_pole(x)
}
\arguments{
\item{x}{data.frame containing the sequence of rotations or the rotation}
}
\value{
Object of class \code{"euler.pole"}
}
\description{
Euler pole on the other side of the hemisphere
}
\examples{
euler.pole <- euler_pole(90, 0)
antipodal_euler_pole(euler.pole)
}
