% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{extract_stage_rotation}
\alias{extract_stage_rotation}
\title{Stage rotation}
\usage{
extract_stage_rotation(r1, r2)
}
\arguments{
\item{r1, r2}{Rotation parameters of two total reconstruction rotations. Both rotations must have the same "fixed" plate.}
}
\value{
list. Rotation is given as the rotation from an older to a younger past position.
}
\description{
Extracts stage rotation between two total reconstruction rotations
}
\details{
\deqn{
R_{[t_2, t_1]} = R_{[t_1, t_2]}^{-1} = R_{[t_0, t_2]} R_{[t_0, t_1]}^{-1}
}
}
\examples{
x <- data.frame(plate.rot = 101, age = c(400, 530), lat = c(72.72, 68.3319), lon = c(150, 11.5198), angle = c(25.59, -31.9464), plate.fix = 301)
extract_stage_rotation(x[1,], x[2, ])
}
\references{
Schaeben, H., Kroner, U. and Stephan, T. (2021). Euler Poles of Tectonic
Plates. In B. S. Daza Sagar, Q. Cheng, J. McKinley and F. Agterberg (Eds.),
*Encyclopedia of Mathematical Geosciences. Encyclopedia of Earth Sciences Series*
(pp. 1--7). Springer Nature Switzerland AG 2021.
\doi{10.1007/978-3-030-26050-7_435-1}
}
\seealso{
\code{\link{extract_stage_rotations}}, \code{\link{extract_stage_rotation_matrix}}
}
