% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{extract_stage_rotation}
\alias{extract_stage_rotation}
\title{Stage rotation extraction}
\usage{
extract_stage_rotation(a1, a2)
}
\arguments{
\item{a1}{rotation matrix of finite rotation 1}

\item{a2}{rotation matrix of finite rotation 2}
}
\value{
list
}
\description{
extract stage rotation of two finite rotations
}
\details{
x must ba all equivalent total rotations.
}
\references{
<div class="csl-entry">Greiner, B. (1999). Euler rotations in plate-tectonic reconstructions. <i>Computers and Geosciences</i>, <i>25</i>(3), 209–216. https://doi.org/10.1016/S0098-3004(98)00160-5</div>
}
\seealso{
\code{\link{extract_stage_rotations}}
}
