% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{finite_pole_interpolation}
\alias{finite_pole_interpolation}
\title{Interpolate gaps in the sequence of total reconstruction rotations}
\usage{
finite_pole_interpolation(rot1, rot2, tx)
}
\arguments{
\item{rot1}{data.frame}

\item{rot2}{data.frame. Must have the same fixed plate as rot1}

\item{tx}{number. Age of the requested intermediate finite rotation
Must be in between \code{rot1$age} and \code{rot2$age}}
}
\value{
data.frame. Sequence of total
reconstruction rotations with filled gaps
}
\description{
Interpolate missing rotations in a sequence of rotations with a
different reference systems
}
\references{
Greiner, B. (1999). Euler rotations in plate-tectonic reconstructions. Computers and Geosciences, 25(3), 209–216. https://doi.org/10.1016/S0098-3004(98)00160-5
}
\seealso{
\code{\link{check.finite}}
}
