% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{grid_points}
\alias{grid_points}
\title{Create a grid}
\usage{
grid_points(gridsize)
}
\arguments{
\item{gridsize}{grid size in degree}
}
\value{
data.frame
}
\description{
Create a grid of equally spaced points
}
