% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{inverse_rotation}
\alias{inverse_rotation}
\title{Invert rotation}
\usage{
inverse_rotation(x)
}
\arguments{
\item{x}{object of class \code{"finite"} or \code{'stage'}}
}
\value{
object of with same class like x
}
\description{
Changes plate motion (A relative to B) to (B relative to A)
}
\seealso{
\code{\link{"as.finite"}}, \code{\link{"as.stage"}}
}
