% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{pole_distribution}
\alias{pole_distribution}
\title{Euler pole statistics}
\usage{
pole_distribution(x)
}
\arguments{
\item{x}{matrix. containing the coordinates of the Euler pole}
}
\description{
Euler pole statistics
}
