% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay.R
\name{ramsay_coords}
\alias{ramsay_coords}
\alias{cartesian_to_spherical}
\alias{spherical_to_cartesian}
\alias{geographical_to_cartesian2}
\alias{cartesian_to_geographical2}
\alias{geographical_to_spherical}
\alias{spherical_to_geographical}
\title{Spherical coordinate conversion}
\usage{
cartesian_to_spherical(x)

spherical_to_cartesian(x)

geographical_to_cartesian2(x)

cartesian_to_geographical2(x)

geographical_to_spherical(x)

spherical_to_geographical(x)
}
\arguments{
\item{x}{matrix}
}
\description{
Spherical coordinate conversion
}
\examples{
x <- rbind(
  c(58, 78, 35),
  c(47, 72, 47),
  c(57, 68, 41),
  c(56, 64, 45),
  c(52, 65, 47),
  c(61, 62, 51),
  c(55, 60, 49),
  c(61, 62, 42),
  c(68, 52, 46),
  c(42, 58, 65)
)
spherical_to_cartesian(x)
spherical_to_geographical(x)
cartesian_to_spherical(spherical_to_cartesian(x))
}
